+++ lib/empathetech_flutter_ui.dart
@@ -7,28 +7,35 @@
 library empathetech_flutter_ui;
 
 // Classes
+export 'src/classes/responsive/EzRow.dart';
+export 'src/classes/responsive/EzRowCol.dart';
+export 'src/classes/responsive/EzSwapScaffold.dart';
+
+export 'src/classes/settings/EzColorSetting.dart';
+export 'src/classes/settings/EzDominantHandSwitch.dart';
+export 'src/classes/settings/EzFontSetting.dart';
+export 'src/classes/settings/EzImageSetting.dart';
+export 'src/classes/settings/EzResetButton.dart';
+export 'src/classes/settings/EzSliderSetting.dart';
+export 'src/classes/settings/EzThemeModeSwitch.dart';
+
+export 'src/classes/text/EzLink.dart';
+export 'src/classes/text/EzPlainText.dart';
+export 'src/classes/text/EzRichText.dart';
+export 'src/classes/text/EzText.dart';
+export 'src/classes/text/EzTextBlock.dart';
+export 'src/classes/text/EzTextStyle.dart';
+export 'src/classes/text/EzTextTheme.dart';
+export 'src/classes/text/EzWebLink.dart';
+
 export 'src/classes/EzAlertDialog.dart';
 export 'src/classes/EzAppProvider.dart';
 export 'src/classes/EzBackAction.dart';
-export 'src/classes/EzColors.dart';
-export 'src/classes/EzColorSetting.dart';
 export 'src/classes/EzConfig.dart';
-export 'src/classes/EzDominantHandSwitch.dart';
-export 'src/classes/EzFontSetting.dart';
 export 'src/classes/EzImage.dart';
-export 'src/classes/EzImageSetting.dart';
-export 'src/classes/EzResetButton.dart';
-export 'src/classes/EzRow.dart';
-export 'src/classes/EzRowCol.dart';
 export 'src/classes/EzScreen.dart';
 export 'src/classes/EzScrollView.dart';
-export 'src/classes/EzSliderSetting.dart';
 export 'src/classes/EzSpacer.dart';
-export 'src/classes/EzSwapScaffold.dart';
-export 'src/classes/EzText.dart';
-export 'src/classes/EzTextStyle.dart';
-export 'src/classes/EzTextTheme.dart';
-export 'src/classes/EzThemeModeSwitch.dart';
 export 'src/classes/EzTransitions.dart';
 export 'src/classes/EzVideoPlayer.dart';
 export 'src/classes/EzWarning.dart';
@@ -36,7 +43,7 @@ export 'src/classes/EzWarning.dart';
 // Constants
 export 'src/consts/empathetechConfig.dart';
 export 'src/consts/googleFonts.dart';
-export 'src/consts/sharedPreferences.dart';
+export 'src/consts/keys.dart';
 
 // Enums
 export 'src/enums/ButtonVis.dart';
@@ -44,12 +51,13 @@ export 'src/enums/Hand.dart';
 export 'src/enums/SliderSettingType.dart';
 
 // Functions
-export 'src/functions/dialogs.dart';
-export 'src/functions/ezAlertActions.dart';
+export 'src/functions/aliases.dart';
+export 'src/functions/colors.dart';
+export 'src/functions/customDialogs.dart';
 export 'src/functions/ezThemeData.dart';
-export 'src/functions/helpers.dart';
 export 'src/functions/navigators.dart';
 export 'src/functions/storage.dart';
+export 'src/functions/text.dart';
 
 // l10n
 export 'src/l10n/efui_phrases.dart';