+++ lib/src/functions/aliases.dart
@@ -1,0 +1,25 @@
+/* empathetech_flutter_ui
+ * Copyright (c) 2023 Empathetech LLC. All rights reserved.
+ * See LICENSE for distribution and usage details.
+ */
+
+import 'package:flutter/material.dart';
+
+/// Do you have a void [Function] as a parameter that you want to be optional?
+/// Then do nothing!
+void doNothing() {}
+
+/// More readable than FocusScope.of(context).unfocus();
+void closeKeyboard(BuildContext context) {
+  FocusScope.of(context).unfocus();
+}
+
+/// More readable than MediaQuery.of(context).size.width
+double widthOf(BuildContext context) {
+  return MediaQuery.of(context).size.width;
+}
+
+/// More readable than MediaQuery.of(context).size.height
+double heightOf(BuildContext context) {
+  return MediaQuery.of(context).size.height;
+}