+++ lib/src/functions/text.dart
@@ -6,25 +6,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-/// Do you have a void [Function] as a parameter that you want to be optional?
-/// Then do nothing!
-void doNothing() {}
-
-/// More readable than FocusScope.of(context).unfocus();
-void closeKeyboard(BuildContext context) {
-  FocusScope.of(context).unfocus();
-}
-
-/// More readable than MediaQuery.of(context).size.width
-double widthOf(BuildContext context) {
-  return MediaQuery.of(context).size.width;
-}
-
-/// More readable than MediaQuery.of(context).size.height
-double heightOf(BuildContext context) {
-  return MediaQuery.of(context).size.height;
-}
-
 /// Returns the soon-to-be rendered size of text via a [TextPainter]
 /// [scalar] should be the value from MediaQuery.of(context).textScaleFactor
 Size measureText({